SUW: Create minimal setup wizard
Brings the Setup Wizard down to the absolute minimum required in order to set up
a device and leave it in a provisioned state (which isn't always possible if the
Setup Wizard is simply omitted). Essentially, this leaves only the following:

- Bluetooth setup (which doesn't require user interaction)
- Welcome screen
- Date/time
- The final page that shows the logo and start button

All of the other settings from Setup Wizard (i.e. Wi-Fi, SIM, etc.) can already
be configured in Settings, so this allows the initial setup to be streamlined
and let the user "get right into it."

Additionally, the accent color has been changed from the hard-coded "LineageOS
teal" to the defined device accent color. This is normally the same thing, but
if a LineageOS-based custom ROM changes this, the Setup Wizard will use it to
maintain a consistent look.

Thanks to XDA user Aqq123, whose work on modifying the Setup Wizard script was
the inspiration for this patch.